refactor(shell): consolidate Open function into shell.go
Move the Open function from process.go to shell.go to simplify the code structure and eliminate redundant files. This change improves maintainability by centralizing shell-related functionality in a single file.
